Update config.md
In fact, an error will be reported if CRF is set to 0 under H264:

```
An error occurred:
TypeError: Setting the CRF to 0 with a H264 codec is not supported anymore because of it's inconsistencies between platforms. Videos with CRF 0 cannot be played on iOS/macOS. 0 is a extreme value with inefficient settings which you probably want. Set CRF to a higher value to fix this error.
```
Former-commit-id: 10baa3c49e0be1b85b616472b67bb3d2c326c5cb