Rendering datasets bundle needs webpack override
Looking over the dataset example by Alex Fernandez, at the issues around Lambda based rendering, and from my own experience just now, it seems that when we're trying to render programmatically, the bundler needs access to the webpack override. If we don't provide this, the video output is blank/black.

Simple documentation change.