remove bucketName input from lambda get-sites
This input is not used in the code.

Not sure if you want to retain the GetSitesInput or put the type declaration inline?
Removing the exported GetSitesInput type would mean removing it from `packages/lambda/src/index.ts` as well. I couldn't see that it was used anywhere else, so it would probably be a good thing to clean up - let me know if you want me to do this.