Docs: Add Google Analytics
So far we had no analytics at all.

We want to understand which use cases people are most interested in and work on those more.

One of our investors has also offered to help us with SEO, and his team is most familiar with Google Analytics, hence why we are going with it.