`@remotion/compositor`: Calculate frame size based on line size
Making a corrupt video behave correctly that had a width of 1718px in it's metadata, but based on the linesize + data size, it only was 1714px wide, causing a SIGSEGV.

Now, we calculate based on the linesize, but the calculation assumes YUY420, the most common pixel format. For other pixel formats, we fall back to previous calculation.

Fixes #5470